+++ src/components.d.ts
@@ -16,8 +16,10 @@ declare module 'vue' {
     ConferenceEvents: typeof import('./components/event-sale/ConferenceEvents.vue')['default']
     ContainerOne: typeof import('./components/homeComponents/ContainerOne.vue')['default']
     CreatedFacilityContainer: typeof import('./components/CreatedFacilityContainer.vue')['default']
+    EventsSlider: typeof import('./components/event-sale/EventsSlider.vue')['default']
     FooterBar: typeof import('./components/FooterBar.vue')['default']
     HighlightEvents: typeof import('./components/event-sale/HighlightEvents.vue')['default']
+    MobileNav: typeof import('./components/home-layout/MobileNav.vue')['default']
     NavBar: typeof import('./components/NavBar.vue')['default']
     Networkingevents: typeof import('./components/event-sale/Networkingevents.vue')['default']
     OrganizationEvent: typeof import('./components/OrganizationEvent.vue')['default']
