+++ src/layouts/home.vue
@@ -1,12 +1,15 @@
 <template>
   <v-app>
     <v-main>
+      <NavBar />
+      <MobileNav/>
       <router-view />
+      <FooterBar />
     </v-main>
-    
   </v-app>
 </template>
 
 <script lang="ts" setup>
-  //
+
+//
 </script>
