+++ src/pages/index.vue
@@ -1,10 +1,7 @@
 <template>
     <div>
-      <NavBar />
-        <v-main>
-          <ContainerOne />
-          <ConatinerTwo />
-        </v-main>
+      <ContainerOne />
+      <ConatinerTwo />
       <FooterBar/>
     </div>
 </template>
@@ -14,3 +11,8 @@ import ContainerOne from "@/components/homeComponents/ContainerOne.vue"
 import ConatinerTwo from "@/components/homeComponents/ConatinerTwo.vue";
 
 </script>
+
+<route lang="yaml">
+  meta :
+    layout : home
+</route>
